requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
tensorflow
transformers
python-multipart
PyPDF2>=3.0.0
keybert
sentence-transformers
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
google-genai
python-dotenv>=1.0.0
peft
keyphrase-vectorizers
nltk
motor
pymongo
python-jose[cryptography]   # JWT signing/verification
passlib[bcrypt]             # Password hashing
email-validator             # For validating user emails in schemas
boto3
httpx
selectolax
lxml
tldextract
tenacity
loguru
langgraph
playwright 
selenium